pyproject.toml
[project]
name = "badi-predictor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "websockets>=12.0",
  "asyncpg>=0.29",
  "tenacity>=8.2",
  "pydantic>=2.0",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pandas>=2.2",
  "xgboost>=2.0",
  "scikit-learn>=1.4",
  "holidays>=0.46",
  "httpx>=0.27",
  "python-dotenv>=1.0",
  "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.3",
  "black>=24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["integration: marks tests as integration tests (deselect with -m 'not integration')"]

[tool.setuptools.packages.find]
include = ["collector*", "ml*", "api*"]

[tool.ruff]
line-length = 88

[tool.black]
line-length = 88
